+++ R/qsc.R
@@ -1,12 +1,17 @@
-#' Title
+#' Quadrilateralized Spherical Cube (QSC)
 #'
-#' @return
+#' The QSC is a set of six equal area projections for each side of the cube. Here
+#' a raw rendition of the cube is returned as six quad primitives in a `mesh3d` object.
+#'
+#' It's not clear if this is useful.
+#' @references \url{https://github.com/OSGeo/proj.4/wiki/Qsc}
+#' @return mesh3d
 #' @export
 #'
 #' @examples
-#' rgl.clear()
+#' library(rgl); #rgl.clear()
 #' wire3d(qsc())
-#' rglwidget()
+#' if ( rgl::rgl.useNULL()) rglwidget()
 qsc <- function() {
  radius <- 6378137
  xy <- rbind(c(-radius, -radius, radius, radius),